// frontend/src/app/model/Product.ts
export interface Product {
    code: string;
    name: string;
    type: string;
    brand: string;
    quantity_type: string;
    rate_per_quantity: string;
    stock_count: string;
    add_date: Date;
    aisle: string;
    shelf: string;
    date_of_manufacture: Date;
    date_of_expiry: Date;
    image: string;
}